requirements.txt
aiogram>=3.4,<4.0
python-dotenv>=1.0.0,<2.0.0